% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R, R/CTDQuery-enrich.R,
%   R/CTDQuery-extract.R, R/CTDQuery-get_terms.R, R/CTDQuery-plot.R
\docType{class}
\name{CTDdata}
\alias{CTDdata}
\alias{CTDdata-class}
\alias{enrich,CTDdata-method}
\alias{get_table,CTDdata-method}
\alias{get_terms,CTDdata-method}
\alias{plot,CTDdata,ANY-method}
\title{Class CTDdata}
\usage{
\S4method{enrich}{CTDdata}(x, y, universe, use = "curated", warnings = TRUE,
  ...)

\S4method{get_table}{CTDdata}(object, index_name, ...)

\S4method{get_terms}{CTDdata}(object)

\S4method{plot}{CTDdata,ANY}(x, y, index_name = "base",
  representation = "heatmap", ...)
}
\arguments{
\item{x}{Object of class \code{CTDdata}.}

\item{y}{NOT USED}

\item{universe}{String vector of genes used as universe. If not provided,
all genes in CTDbase are used.}

\item{use}{Select if all or only curated relations are used.}

\item{warnings}{Shows or hiddes warnings.}

\item{...}{NOT USED}

\item{object}{Object of class \code{CTDdata}.}

\item{index_name}{Name of the plot to be draw. \code{"base"} shows a
bar-plot indicating the number of terms lost&found at CTDbase. For
\emph{gene} queries \code{index_name} can take values \code{"disease"},
\code{"chemical"}, \code{"gene-gene interaction"}, \code{"kegg pathways"}
and \code{"go terms"}. The first two have only \code{"heatmap"}
representation, the last to have only \code{"network"} representation, and
\code{"gene-gene interaction"} has both. For \emph{chemical} queries
\code{index_name} can take values \code{"gene"}, \code{"disease"},
\code{"kegg pathways"} and \code{"go terms"}. The first has both
\code{"heatmap"} and \code{"network"} representation, while the last only
\code{"network"}. The otehr three have only \code{"heatmap"} representation.
For \emph{disease} queries \code{index_name} can take values \code{"gene"},
\code{"chemical"} and \code{"kegg pathways"}. The first two have
\code{"heatmap"} representation while the last \code{"network"}
representation.}

\item{representation}{Can take values \code{"heatmap"} or \code{"network"}.}
}
\value{
An object of class \code{CTDdata}
}
\description{
Class resulting of \code{\link{query_ctd_gene}},
\code{\link{query_ctd_chem}} and \code{\link{query_ctd_dise}}. It is used
to encapsulate all the information in \emph{CTDbase} for given set of genes,
chemicals or diseases.
}
\section{Methods (by generic)}{
\itemize{
\item \code{enrich}: Method to perform enrichment analysis given two
object of class \code{CTDdata}.

\item \code{get_table}: Method to obtain a specific inner table from an \code{CTDdata} object.

\item \code{get_terms}: Return a list with the terms found to create the object.

\item \code{plot}: Generates a basic plot showing the number of terms
that can be used to query CTDbase.
}}

\section{Slots}{

\describe{
\item{\code{type}}{Character saving \code{"GENE"}, \code{"CHEMICAL"} or
\code{"DISEASE"} depending if it was created using
\code{\link{query_ctd_gene}}, \code{\link{query_ctd_chem}} or
\code{\link{query_ctd_dise}}}

\item{\code{terms}}{\code{DataFrame} with the genes, chemicals or diseases used
to create the object.}

\item{\code{losts}}{Character with the terms used to create the object
but that were nor present in CTDbase.}

\item{\code{gene_interactions}}{(Only for chemicals) Table with a relation
of the genes interacting with the given chemicals.}

\item{\code{chemicals_interactions}}{(Only for genes) Table with a relation
of the chemicals interacting with the given genes.}

\item{\code{diseases}}{Table with a relation of the diseases associated with
given genes or chemicals.}

\item{\code{gene_gene_interactions}}{(Only for genes) Table with a relation
of the genes interacting with the given genes.}

\item{\code{kegg}}{Table with a relation of the KEGG pathways affected by the
given chemicals or where the given genes play a role.}

\item{\code{go}}{Table with a relatio of the GO terms afected by the given
chemicals or where the given genes play a role.}
}}

\seealso{
\code{\link{query_ctd_gene}} to create a \code{CTDdata}
from a set of genes, \code{\link{query_ctd_chem}} to create a
\code{CTDdata} from a set of chemicals, \code{\link{query_ctd_dise}} to
create a \code{CTDdata} from a set of diseases,
\code{\link{get_table}} to retrive encapsulated data and
\code{plot} to get nice plots from stored data.
}
