% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTDquerier.R
\docType{package}
\name{CTDquerier}
\alias{CTDquerier}
\title{Package for CTDbase data query, data visualization and data analysis.}
\description{
It can retrieve information related to genes, chemicals and diseases.
}
\section{Data Download}{

\code{CTDquerier} offers two functions to query CTDbase (http://ctdbase.org):
\code{\link{query_ctd_gene}} to query CTDbase given a set of genes; and
\code{\link{query_ctd_chem}} to query CTDbase given a set of chemicals. Both
functions returns \code{\link{CTDdata}} objects. Raw downloaded information
can be retrieved from \code{\link{CTDdata}} using method
\code{\link{get_table}}.
}

\section{Data Visualization}{

\code{\link{CTDdata}} objects offers basic visualization of the downloaded
information using standard \code{plot} method.
}

