% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R, R/CTDQuery-enrich.R,
%   R/CTDQuery-extract.R, R/CTDQuery-get_terms.R, R/CTDQuery-plot.R
\docType{class}
\name{CTDquery}
\alias{CTDquery}
\alias{CTDquery-class}
\alias{enrich,CTDquery-method}
\alias{extract,CTDquery-method}
\alias{get_terms,CTDquery-method}
\alias{plot,CTDquery,ANY-method}
\title{Class CTDquery}
\usage{
\S4method{enrich}{CTDquery}(x, y, ...)

\S4method{extract}{CTDquery}(object, index_name, ...)

\S4method{get_terms}{CTDquery}(object)

\S4method{plot}{CTDquery,ANY}(x, y, index_name = "base",
  representation = "heatmap", ...)
}
\arguments{
\item{x}{Object of class \code{CTDquery}.}

\item{y}{NOT USED}

\item{...}{NOT USED}

\item{object}{Object of class \code{CTDquery}.}

\item{index_name}{Name of the plot to be draw. \code{"base"} shows a
bar-plot indicating the number of terms lost&found at CTDbase. For
\emph{gene} queries \code{index_name} can take values \code{"disease"},
\code{"chemical"}, \code{"gene-gene interaction"}, \code{"kegg pathways"}
and \code{"go terms"}. The first two have only \code{"heatmap"}
representation, the last to have only \code{"network"} representation, and
\code{"gene-gene interaction"} has both. For \emph{chemical} queries
\code{index_name} can take values \code{"gene"}, \code{"disease"},
\code{"kegg pathways"} and \code{"go terms"}. The first has both
\code{"heatmap"} and \code{"network"} representation, while the last only
\code{"network"}. The otehr three have only \code{"heatmap"} representation.
For \emph{disease} queries \code{index_name} can take values \code{"gene"},
\code{"chemical"} and \code{"kegg pathways"}. The first two have
\code{"heatmap"} representation while the last \code{"network"}
representation.}

\item{representation}{Can take values \code{"heatmap"} or \code{"network"}.}
}
\value{
An object of class \code{CTDquery}
}
\description{
Class resulting of \link{\code{query_ctd_gene}},
\link{\code{query_ctd_chem}} and \link{\code{query_ctd_dise}}. It is used
to encapsulate all the information in \emp{CTDbase} for given set of genes,
chemicals or diseases.
}
\details{
Available tables are (\code{index_name}):
\enumerate{
  \item \code{"genes interactions"}: (Only for chemicals) Table with a relation
  of the genes interacting with the given chemicals.
  \item \code{"chemicals interactions"}: (Only for genes) Table with a relation
  of the chemicals interacting with the given genes.
  \item \code{"diseases"}: Table with a relation of the diseases associated with
  given genes or chemicals.
  \item \code{"gene-gene interactions"}: (Only for genes) Table with a relation
  of the genes interacting with the given genes.
  \item \code{"kegg pathways"}: Table with a relation of the KEGG pathways affected by the
given chemicals or where the given genes play a role.
  \item \code{"go terms"}: Table with a relatio of the GO terms afected by the given
chemicals or where the given genes play a role.
}
}
\section{Methods (by generic)}{
\itemize{
\item \code{enrich}: MEthod to perform enrichment analysis given two
object of class \code{CTDquery}.

\item \code{extract}: Method to obtain a specific inner table from an \code{CTDquery} object.

\item \code{get_terms}: Return a list with the terms found to create the object.

\item \code{plot}: Generates a basic plot showing the number of terms
that can be used to query CTDbase.
}}

\section{Slots}{

\describe{
\item{\code{type}}{Character saving \code{"GENE"}, \code{"CHEMICAL"} or
\code{"DISEASE"} depending if it was created using
\code{\link{prepare_ctd_gene}}, \code{\link{prepare_ctd_chem}} or
\link{\code{query_ctd_dise}}}

\item{\code{terms}}{\code{DataFrame} with the genes, chemicals or diseases used
to create the object.}

\item{\code{losts}}{Character with the terms used to create the object
but that were nor present in CTDbase.}

\item{\code{gene_interactions}}{(Only for chemicals) Table with a relation
of the genes interacting with the given chemicals.}

\item{\code{chemicals_interactions}}{(Only for genes) Table with a relation
of the chemicals interacting with the given genes.}

\item{\code{diseases}}{Table with a relation of the diseases associated with
given genes or chemicals.}

\item{\code{gene_gene_interactions}}{(Only for genes) Table with a relation
of the genes interacting with the given genes.}

\item{\code{kegg}}{Table with a relation of the KEGG pathways affected by the
given chemicals or where the given genes play a role.}

\item{\code{go}}{Table with a relatio of the GO terms afected by the given
chemicals or where the given genes play a role.}
}}

\note{
The pairs of \code{CTDquery} objects given to x and y must follow a
specific pattern: gene-disease, gene-chemical or disease-chemical.
}
\seealso{
\code{\link{query_ctd_gene}} to create a \code{CTDquery}
from a set of genes, \code{\link{query_ctd_chem}} to create a
\code{CTDquery} from a set of chemicals, \link{\code{query_ctd_dise}} to
create a \code{CTDquery} from a set of diseases,
\code{\link[psygenet2r]{extract}} to retrive encapsulated data and
\code{plot} to get nice plots from stored data.
}
