% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdStorage.R
\docType{class}
\name{ctdStorage}
\alias{ctdStorage}
\title{Database storage from CTDbase}
\description{
Database storage from CTDbase

Database storage from CTDbase
}
\details{
Encapsulation of one or multiple database files from _CTDbase_.
}
\keyword{association,}
\keyword{ctdbase}
\keyword{data,}
\keyword{database,}
\keyword{interaction,}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ctdStorage$new()}}
\item \href{#method-print}{\code{ctdStorage$print()}}
\item \href{#method-includes}{\code{ctdStorage$includes()}}
\item \href{#method-clone}{\code{ctdStorage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new `ctdStorage` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctdStorage$new(file_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_list}}{List of character vectors. Each vector containing two elements: name and path.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ctdStorage` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Shows a brief description of the `ctdStorage`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctdStorage$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Do not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `invisible` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-includes"></a>}}
\if{latex}{\out{\hypertarget{method-includes}{}}}
\subsection{Method \code{includes()}}{
Indicates if a given key is present in the `ctdStorage`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctdStorage$includes(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{`character` with the key to be checked.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `boolean` vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctdStorage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
