% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdVocabulary.R
\docType{class}
\name{ctdTerm}
\alias{ctdTerm}
\title{Term from CTDbase's Vocabulary}
\description{
Term from CTDbase's Vocabulary

Term from CTDbase's Vocabulary
}
\details{
Result of validating one or more terms in a `ctdVocabulary` object.
}
\keyword{ctdbase}
\keyword{data,}
\keyword{term,}
\keyword{vocabulary,}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ctdTerm$new()}}
\item \href{#method-get_terms}{\code{ctdTerm$get_terms()}}
\item \href{#method-print}{\code{ctdTerm$print()}}
\item \href{#method-clone}{\code{ctdTerm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new `ctdTerm` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctdTerm$new(data_source, user_terms, validated_terms)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_source}}{Source of data of the file to be loaded, one of: _gene_, _disease_, _chemical_, or _pathway_.}

\item{\code{user_terms}}{`character` with all the terms given by user to be validated.}

\item{\code{validated_terms}}{`data.table` with all the information related to terms in _CTDbase_.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ctdTerm` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_terms"></a>}}
\if{latex}{\out{\hypertarget{method-get_terms}{}}}
\subsection{Method \code{get_terms()}}{
Returns the terms used to create the `ctdTerm` indicating which of them where validated.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctdTerm$get_terms()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A `list` with two entries classifying the terms (`valid`, `no_valid`) according if they are in _CTDbase_.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Shows a brief description of the `ctdTerm`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctdTerm$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Do not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `invisible` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctdTerm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
