% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdVocabulary.R
\docType{class}
\name{ctdVocabulary}
\alias{ctdVocabulary}
\title{Vocabulary from CTDbase}
\description{
Vocabulary from CTDbase

Vocabulary from CTDbase
}
\details{
Encapsulation of a _CTDbase_ vocabulary file.
}
\keyword{ctdbase}
\keyword{data,}
\keyword{vocabulary,}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ctdVocabulary$new()}}
\item \href{#method-source}{\code{ctdVocabulary$source()}}
\item \href{#method-table}{\code{ctdVocabulary$table()}}
\item \href{#method-print}{\code{ctdVocabulary$print()}}
\item \href{#method-search_term}{\code{ctdVocabulary$search_term()}}
\item \href{#method-validate_term}{\code{ctdVocabulary$validate_term()}}
\item \href{#method-clone}{\code{ctdVocabulary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new `ctdVocabulary` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctdVocabulary$new(file_path, data_source, data_format = "csv")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_path}}{Path, including the file's name, to the file to be loaded.}

\item{\code{data_source}}{Source of data of the file to be loaded, one of: _gene_, _disease_, _chemical_, or _pathway_.}

\item{\code{data_format}}{Format of the file to be loaded (`"csv"` for `"csv"` or `"csv.gz"` file, or `"tsv"` for `"tsv"` or `"tsv.gz"` file).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ctdVocabulary` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-source"></a>}}
\if{latex}{\out{\hypertarget{method-source}{}}}
\subsection{Method \code{source()}}{
Returns a `character` indicating the _source_ of the `ctdVocabulary`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctdVocabulary$source()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A `character` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-table"></a>}}
\if{latex}{\out{\hypertarget{method-table}{}}}
\subsection{Method \code{table()}}{
Returns the content of the `ctdVocabulary`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctdVocabulary$table()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A `data.table` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Shows a brief description of the `ctdVocabulary`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctdVocabulary$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Do not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `invisible` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search_term"></a>}}
\if{latex}{\out{\hypertarget{method-search_term}{}}}
\subsection{Method \code{search_term()}}{
Searches for a term in a given `ctdVocabulary`. The following fields can be used for the search:
  * Genes: _gene symbol_, _gene name_, and _gene id_.
  * Diseases: _disease name_ and _disease id_.
  * Chemicals: _chemical name_, _chemical id_, and _CAS registry number_.
  * Pathway: _pathway name_ and _pathway id_.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctdVocabulary$search_term(terms)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{terms}}{Terms to search in the given `ctdVocabulary` object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `data.table` object. First column is filled with user term, second with internal "name", and the third with the similarity score, computed using _longest common substring distance_.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_term"></a>}}
\if{latex}{\out{\hypertarget{method-validate_term}{}}}
\subsection{Method \code{validate_term()}}{
Validates a series of terms into the `ctdVocabulary`.The following fields are used for validating the terms:
  * Genes: _gene symbol_, _gene name_, and _gene id_.
  * Diseases: _disease name_ and _disease id_.
  * Chemicals: _chemical name_, _chemical id_, and _CAS registry number_.
  * Pathway: _pathway name_ and _pathway id_.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctdVocabulary$validate_term(terms)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{terms}}{Terms to validate in the given `ctdVocabulary` object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `ctdTerm` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctdVocabulary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
