% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ctd.R
\name{download_ctd_genes}
\alias{download_ctd_genes}
\title{Function to download genes available in CTDbase}
\usage{
download_ctd_genes(filename = "CTD_genes.tsv.gz", mode = "auto",
  verbose = FALSE)
}
\arguments{
\item{filename}{(default \code{"CTD_genes.tsv.gz"}) Name of the file in
the local system.}

\item{mode}{(default \code{"auto"}) Mode passed to \code{download.file}.}
}
\description{
This function download the "Gene vocabulary" file (\code{CTD_genes.tsv.gz})
from \code{http://ctdbase.org/downloads}.
}
\details{
The field included in the file (\code{CTD_genes.tsv.gz}) are:
\enumerate{
 \item GeneSymbol
 \item GeneName
 \item GeneID (NCBI Gene identifier)
 \item AltGeneIDs (alternative NCBI Gene identifiers; '|'-delimited list)
 \item Synonyms ('|'-delimited list)
 \item BioGRIDIDs ('|'-delimited list)
 \item PharmGKBIDs ('|'-delimited list)
 \item UniprotIDs ('|'-delimited list)
}
}
\examples{
download_ctd_genes()
file.exists( "CTD_genes.tsv.gz" )
}
