% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllGenerics.R
\name{enrich}
\alias{enrich}
\title{Method to perform enrichment analysis using two \code{CTDdata} objects}
\usage{
enrich(x, y, universe, use = "curated", warnings = TRUE, ...)
}
\arguments{
\item{x}{Object of class \code{\link{CTDdata}}.}

\item{y}{Object of class \code{\link{CTDdata}}.}

\item{universe}{Vector of strings corresponding to the genes to be used
as universe.}

\item{use}{(default: \code{"curated"}) It can take values \code{"curated"}
or \code{"all"} to filter or not filter for curated the genes into
the \code{CTDdata} objects.
and \code{phyper} (with \code{lower.tail=TRUE}) respectively.}

\item{warnings}{(default: \code{TRUE}).}

\item{...}{NOT USED}
}
\value{
A list with class \code{htest}. Check
\code{fisher.test} for more information.
}
\description{
This methods performs a fisher test using the genes in two objects of
class \code{\link{CTDdata}}. The object in 'x' is used as source while
the object on 'y' is used as universe. When object 'x' corresponds to
an object created with \code{\link{query_ctd_gene}}, the used genes
are the found terms in CTDbase. In the other cases (chemicals and
disease \code{\link{CTDdata}}), the genes from  the 'gene interactions'
table are used. If \code{universe} is missing, all genes in CTDbase
are used as universe.
}
