% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gala.R
\docType{data}
\name{gala}
\alias{gala}
\title{\code{CTDdata} for ilustrative purpouses}
\format{
An object of class \code{CTDdata} of length 1.
}
\usage{
data("gala")
}
\value{
An \code{CTDdata} object.
}
\description{
\code{CTDdata} with information of 258 genes downloaded from CTDbase.
 The object was created from from the genes obtained from  the scientific
 article entitleed Case-control admixture mapping in Latino populations
 enriches for known asthma-associated genes" (Table E1) by Torgerson et. al.
 The genes were used to query CTDbase using \code{query_ctd_genes} function.
}
\examples{
data("gala")
gala
}
\keyword{datasets}
