% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctd_url.R
\name{get_ctd_url}
\alias{get_ctd_url}
\title{Given an index, a term and a category it returns the URL from CTD to
download the corresponding information.}
\usage{
get_ctd_url(index, term = "", category = "")
}
\description{
Given an index, a term and a category it returns the URL from CTD to
download the corresponding information.
}
