% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ctd.R
\name{load_ctd_chem}
\alias{load_ctd_chem}
\title{Function to load the \code{.tsv.gz} file for genes}
\usage{
load_ctd_chem(filename = "CTD_chemicals.tsv.gz")
}
\arguments{
\item{filename}{(default "CTD_chemicals.tsv.gz") File with the
chemicals downloaded from CTDbase.}
}
\value{
A \code{data.frame} with the content of the file "CTD_genes.tsv.gz"
}
\description{
Function to load the \code{.tsv.gz} file for genes
}
\details{
The field included in the file (\code{CTD_chemicals.tsv.gz}) are:
\enumerate{
 \item ChemicalName
 \item ChemicalID (MeSH identifier)
 \item CasRN (CAS Registry Number, if available)
 \item Definition
 \item ParentIDs (identifiers of the parent terms; '|'-delimited list),
 \item TreeNumbers (identifiers of the chemical's nodes; '|'-delimited list),
 \item ParentTreeNumbers (identifiers of the parent nodes; '|'-delimited list),
 \item Synonyms ('|'-delimited list)
 \item DrugBankIDs ('|'-delimited list)
 }
}
\examples{
download_ctd_chem()
fdl <- load_ctd_chem()
dim(fdl)
}
