% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ctd_chem.R
\name{query_ctd_chem}
\alias{query_ctd_chem}
\title{Function to query CTDbase using chemical terminology ( Chemical Name )}
\usage{
query_ctd_chem(terms, filename = "CTD_chemicals.tsv.gz", mode = "auto",
  max.distance = 10, verbose = FALSE)
}
\arguments{
\item{terms}{Character vector with the chemicals used in the query.}

\item{filename}{(default \code{"CTD_chemicals.tsv.gz"}) Name of the file
to store the CTDbase chemicals vocabilary.}

\item{mode}{(default \code{"auto"}) Mode passed to \code{download.file}.}

\item{max.distance}{(default \code{10}) Maximum distance allowed between a given element
in \code{terms} argument and a possible match in CTDbase.}

\item{verbose}{(default \code{FALSE}) If set to \code{TRUE} is shows relevant
information of each step.}
}
\value{
An object of class \code{\link{CTDdata}}.
}
\description{
This function cheks for CTDbase gene vocabulary and query CTDbase
for each one, downloading chemica-genes interactions, associated
diseases, associated KEGG pathways and associated GO terms.
}
\examples{
rst <- query_ctd_chem( terms = c( "Iron", "Air Pollutants" ), verbose = TRUE )
}
