% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ctd_dise.R
\name{query_ctd_dise}
\alias{query_ctd_dise}
\title{Function to query CTDbase using disease terminology}
\usage{
query_ctd_dise(terms, filename = "CTD_diseases.tsv.gz", mode = "auto",
  verbose = FALSE)
}
\arguments{
\item{terms}{Character vector with the diseases used in the query.}

\item{filename}{(default \code{"CTD_diseases.tsv.gz"}) Name of the file
to store the CTDbase disease vocabilary.}

\item{mode}{(default \code{"auto"}) Mode passed to \code{download.file}.}

\item{verbose}{(default \code{FALSE}) If set to \code{TRUE} is shows relevant
information of each step.}
}
\value{
An object of class \code{\link{CTDdata}}.
}
\description{
This function cheks for CTDbase disease vocabulary and query CTDbase
for each one, downloading disease-gene interactions, chemicals
interactions, associated desease, associated KEGG pathways and
associated GO terms.
}
\examples{
rst <- query_ctd_dise( terms = "Asthma", verbose = TRUE )
}
