% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTDquerier.R
\docType{package}
\name{CTDquerier}
\alias{CTDquerier}
\alias{CTDquerier-package}
\title{Package for CTDbase data query, data visualization and data analysis.}
\description{
Package for CTDbase data query, data visualization and data analysis.
}
\section{Data Download}{

\code{CTDquerier} offers two functions to query CTDbase (http://ctdbase.org):
\link{\code{query_ctd_gene}} to query CTDbase given a set of genes; and
\link{\code{query_ctd_chem}} to query CTDbase given a set of chemicals. Both
functions returns \link{\code{CTDquery}} objects. Raw downloaded information
can be retrieved from \link{\code{CTDquery}} using method
\code{\link[psygenet2r]{extract}}.
}

\section{Data Visualization}{

\link{\code{CTDquery}} objects offers basic visualization of the downloaded
information using standard \code{plot} method.
}

