% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R, R/CTDQuery-extract.R,
%   R/CTDQuery-plot.R
\docType{class}
\name{CTDquery}
\alias{CTDquery}
\alias{CTDquery-class}
\alias{extract,CTDquery-method}
\alias{plot,CTDquery,ANY-method}
\title{Class CTDquery}
\usage{
\S4method{extract}{CTDquery}(object, name, ....)

\S4method{plot}{CTDquery,ANY}(x, index_name, representation = "heatmap", ...)
}
\arguments{
\item{name}{Name of the table to be retrived.}

\item{x}{Object of class \code{\link{CTDquery}}.}

\item{index_name}{Name of the table to be ploted (see \code{\link{extract}}).}

\item{representation}{Some of the tables can have both network or heat-map
representations.}

\item{...}{Can take anyone of the following values: \code{filter.score} (default
\code{20}), \code{max.length} (default \code{50}, \code{30} or \code{2}),
\code{field.score} (default \code{"Inference"}, can also take \code{"Reference"}),
\code{subset.x} to subset X-axis and \code{subset.y} to subset Y-axis.}
}
\value{
An object of class \code{CTDquery}
}
\description{
Class resulting of \link{\code{prepare_ctd_gene}} and
\link{\code{prepare_ctd_chem}}. It is sued to encapsulate
all the information in CTDbase for given set of genes
or chemicals.
}
\details{
Available tables are:
\enumerate{
  \item \code{"genes interactions"}: (Only for chemicals) Table with a relation
  of the genes interacting with the given chemicals.
  \item \code{"chemicals interactions"}: (Only for genes) Table with a relation
  of the chemicals interacting with the given genes.
  \item \code{"diseases"}: Table with a relation of the diseases associated with
  given genes or chemicals.
  \item \code{"gene-gene interactions"}: (Only for genes) Table with a relation
  of the genes interacting with the given genes.
  \item \code{"kegg pathways"}: Table with a relation of the KEGG pathways affected by the
given chemicals or where the given genes play a role.
  \item \code{"go terms"}: Table with a relatio of the GO terms afected by the given
chemicals or where the given genes play a role.
}
}
\section{Methods (by generic)}{
\itemize{
\item \code{extract}: Method to obtain a specific inner table from an \code{CTDquery} object.

\item \code{plot}: Method to represent CTDbase information of a given query.
}}

\section{Slots}{

\describe{
\item{\code{type}}{Character saving \code{"GENE"} or \code{"CHEMICAL"}
depending if it was created using \code{\link{prepare_ctd_gene}}
or \code{\link{prepare_ctd_chem}}.}

\item{\code{terms}}{\code{DataFrame} with the genes or chemicals used
to create the object.}

\item{\code{losts}}{Character with the terms used to create the object
but that were nor present in CTDbase.}

\item{\code{gene_interactions}}{(Only for chemicals) Table with a relation
of the genes interacting with the given chemicals.}

\item{\code{chemicals_interactions}}{(Only for genes) Table with a relation
of the chemicals interacting with the given genes.}

\item{\code{diseases}}{Table with a relation of the diseases associated with
given genes or chemicals.}

\item{\code{gene_gene_interactions}}{(Only for genes) Table with a relation
of the genes interacting with the given genes.}

\item{\code{kegg}}{Table with a relation of the KEGG pathways affected by the
given chemicals or where the given genes play a role.}

\item{\code{go}}{Table with a relatio of the GO terms afected by the given
chemicals or where the given genes play a role.}
}}

\seealso{
\code{\link{prepare_ctd_gene}} to create a \code{CTDquery}
from a set of genes, \code{\link{prepare_ctd_chem}} to create a
\code{CTDquery} from a set of chemicals and \link{extract} to
retrive encapsulated data.
}
