% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllGenerics.R
\name{enrich}
\alias{enrich}
\title{Method to perform enrichment analysis using two \code{CTDdata} objects}
\usage{
enrich(x, y, ...)
}
\arguments{
\item{x}{Object of class \code{\link{CTDdata}}.}

\item{y}{Object of class \code{\link{CTDdata}}.}

\item{...}{NOT USED}
}
\value{
A list with class \code{htest}. Check
\code{fisher.test} for more information.
}
\description{
This methods performes a fisher test using the genes or chemicals
in two objects of class \code{\link{CTDdata}}.
}
